you need to migrate db

Related work items: #378

+++ app/Services/Impl/ResponseService.php
@@ -108,7 +108,7 @@ class ResponseService extends BaseService implements IResponseService
      *
      * @return Model
      */
-    public function assignToClient($client, $questionnaire)
+    public function assignToClient($client, $questionnaire,$assignfromgroup = false)
     {
         $assigned = $this
             ->optional()
@@ -116,16 +116,19 @@ class ResponseService extends BaseService implements IResponseService
                 ['questionnaire_id', $questionnaire],
                 ['user_id', $client],
             ]);
+		if ($assignfromgroup == true && ! is_null($assigned)) {
+				return false;
+		} else {
+			// Prevent assignment of a questionnaire if it is already assigned
+			if (! is_null($assigned)) {
+				throw new QuestionnaireAlreadyAssignedException();
+			}
 
-        // Prevent assignment of a questionnaire if it is already assigned
-        if (! is_null($assigned)) {
-            throw new QuestionnaireAlreadyAssignedException();
-        }
-
-        return $this->create([
-            'questionnaire_id' => $questionnaire,
-            'user_id' => $client,
-        ]);
+			return $this->create([
+				'questionnaire_id' => $questionnaire,
+				'user_id' => $client,
+			]);
+		}
     }
 
     /**
@@ -238,17 +241,20 @@ class ResponseService extends BaseService implements IResponseService
      *
      * @return int
      */
-    public function unassignFromClient($client, $questionnaire)
+    public function unassignFromClient($client, $questionnaire,$assignfromgroup = false)
     {
         $response = $this->findBy([
             ['questionnaire_id', $questionnaire],
             ['user_id', $client],
         ]);
+		if ($assignfromgroup == true && $response->complete) {
+				return false;
+		} else {
+			if ($response->complete) {
+				throw new QuestionnaireAlreadyCompletedException();
+			}
 
-        if ($response->complete) {
-            throw new QuestionnaireAlreadyCompletedException();
-        }
-
-        return $this->delete($response->id);
+			return $this->delete($response->id);
+		}
     }
 }
